First commit of Dashboard and closet(wardrope)
